tweaked RegEx_intro.md
Bruno's lesson was originally designed to work with regexpal.com, but I prefer regexr.com, so I tweaked the instructions to work better with this website. I also added a challenge question to intorduce escape characters, simplified the email challenge question, and added a second correct answer for challenge question 5.